command changes
